auth-service/requirements.txt
django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
